refactor: library management code into separate class
